Moving to the next Question!!!!
Progress report, questions advance to the next question without issue
and reduce the budget amount and the budget amount is displaying
properly as money.

Current project is to at least create UI for all of the scenarios in the
game...

Need to work on scenario randomization, retention of data, camera
animations.

This is not as straight forward as I thought it would be however I am
starting to figure it out. I mean I think this has the potential to be
good, but I am a little worried about only having about 3 weeks left.